update to use vector ops to find midpoint
